Context: -------------------
"Mahasen AI v013"

We believe life shouldn't be consumed by endless notifications and digital noise. Our mission is to build your perfect digital twin. An AI that knows you so well it can act on your behalf, shielding you from distractions and giving you back control of your time and attention. Mahasen AI is designed to respond as you would, handle routine communications, and only bring truly important matters to your attention. It should be able to see and take action without the user having to set up APIs for every application. The user setup should be as smooth and frictionless as possible.

To do this, I plan to build an auto replying email agent. First it will be a web interface, then I'll make a chrome extension after that. You'll have to build the first email agent with the interface.

For the interface, It should have a dashboard with a welcome message on top. And a sidebar should be there. Menu Items are = [ Dashboard, Mail, Intergrations, Personal Context, Settings ]

This should be a typical chat conversation section, where the user can chat with the AI. The AI should use the personal context when replying to messages. If the integrations of Gmail, calendar, google drive is not completed, it should say so as well.

This section is a complete email inbox of gmail. All the emails from the user's gmail should be fetched here and displayed with little sidebar items in it [ Inbox, Unread, Sent, Drafts.. ]

When I open a letter, it should open just like in gmail. And at the bottom there should be a "Generate Draft" button. When clicked on the Generate Draft button, it should generate 3 draft replies for those emails using the personal context.

Those 3 drafts should be displayed side by side. When clicked on a draft, that draft should get automatically pasted into the bottom reply input email section and I should have the ability to change the now generated & pasted draft. There should be two button under it. "Regenerate Drafts" & "Send Reply"

Regenerate Drafts should regenerate the 3 draft replies and Send reply button should send the email reply.

There should be a onboarding task list summary in this section. Whether the user have given access to their gmail using OAuth. For all, Gmail, Google Drive, Calendar. And if Mahasen is trained to find the personal context of the user. If it's not done, there should be a beautiful "Learn about me from my Email History" button. When clicked on it it should start the "Learn About Human Process"

A section where the user can see their personal profile, Email, name.

This is the section where we can display everything the agent have learned about the user. There should be these information automatically learned from the "Learn About Human Process". But those things should be manually editable too. And when the AI agent is talking using the chatbox in the dashboard, and when it's reading a new email, if it need to change anything in the personal context memory, the agent should have those permission to do so.

How the human reply to certain emails or people or event may change, for an example he may reply in a friendly tone to a friend and in a formal tone to a work email. So these things must be adjusted to account for that as well.

--------------------------------

Your Task:
Now please make the needed changes to complete the Mail section. Complete the Mail Section